requirements.txt
Pillow
selenium
requests
bs4